+++ src/app/admin/components/admin-product-form/admin-product-form.component.ts
@@ -1,12 +1,14 @@
-import { Component, Inject, OnInit } from '@angular/core';
-import { ActivatedRoute, Router, UrlTree } from '@angular/router';
-import { Observable, Subscription } from 'rxjs';
-import { pluck } from 'rxjs/operators';
+import { Component, Inject, OnDestroy, OnInit, ViewChild } from '@angular/core';
+import { NgForm } from '@angular/forms';
+import { UrlTree } from '@angular/router';
+import { Observable, Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+import { ProductsFacade } from 'src/app/core/@ngrx/products';
+import { RouterFacade } from 'src/app/core/@ngrx/router';
 import { CanComponentDeactivate } from 'src/app/core/interfaces/can-component-deactivate.interface';
 import { DialogService } from 'src/app/core/services/dialog.service';
 import { generatedString, GeneratorFactory, GeneratorService } from 'src/app/core/services/generator.service';
 import { Product } from 'src/app/products/models/product.model';
-import { ProductsService } from 'src/app/products/services/products.service';
 
 @Component({
   selector: 'app-admin-product-form',
@@ -17,49 +19,65 @@ import { ProductsService } from 'src/app/products/services/products.service';
     { provide: generatedString, useFactory: GeneratorFactory(10), deps: [GeneratorService] },
   ]
 })
-export class AdminProductFormComponent implements OnInit, CanComponentDeactivate {
+export class AdminProductFormComponent implements OnInit, OnDestroy, CanComponentDeactivate {
   product: Product;
-  originalProduct: Product;
-  private sub: Subscription;
+  private componentDestroyed$: Subject<void> = new Subject<void>();
+  @ViewChild('form', { static: false })
+  userForm: NgForm;
+  private isSubmitClick = false;
 
   constructor(
     @Inject(generatedString) public newID: string,
-    private productsService: ProductsService,
-    private route: ActivatedRoute,
-    private router: Router,
+    private productsFacade: ProductsFacade,
+    private routerFacade: RouterFacade,
     private dialogService: DialogService
   ) { }
 
   ngOnInit(): void {
-    this.route.data.pipe(pluck('product')).subscribe((product: Product) => {
-      this.product = { ...product };
-      this.originalProduct = { ...product };
-    });
+    const observer: any = {
+      next: (product: Product): void => {
+        this.product = {...product};
+      },
+      error(err: Error | string): void {
+        console.log(err);
+      },
+      complete(): void {
+        console.log('Stream is completed');
+      }
+    };
+
+    this.productsFacade.selectedProductByUrl$
+      .pipe(
+        takeUntil(this.componentDestroyed$)
+      )
+      .subscribe(observer);
+  }
+
+  ngOnDestroy(): void {
+    this.componentDestroyed$.next();
+    this.componentDestroyed$.complete();
   }
 
   onSaveProduct(): void {
     const product = { ...this.product };
-
-    const observer = {
-      next: (savedProduct: Product) => {
-        this.originalProduct = { ...savedProduct };
-        this.onGoBack();
-      },
-      error: (err: any) => console.log(err)
-    };
+    this.isSubmitClick = true;
 
     if (product.id) {
-      this.sub = this.productsService.updateProduct(product).subscribe(observer);
+      this.productsFacade.updateProduct({ product });
     } else {
-      this.sub = this.productsService.createProduct({
-        ...product,
-        id: this.newID
-      }).subscribe(observer);
+      this.productsFacade.createProduct({
+        product: {
+          ...product,
+          id: this.newID
+        }
+      });
     }
+
+    this.onGoBack();
   }
 
   onGoBack(): void {
-    this.router.navigate(['/admin/products']);
+    this.routerFacade.goTo({ path: ['/admin/products'] });
   }
 
   canDeactivate():
@@ -67,14 +85,12 @@ export class AdminProductFormComponent implements OnInit, CanComponentDeactivate
     | Promise<boolean | UrlTree>
     | boolean
     | UrlTree {
-    const flags = Object.keys(this.originalProduct).map(key => {
-      if (this.originalProduct[key] === this.product[key]) {
-        return true;
-      }
-      return false;
-    });
 
-    if (flags.every(el => el)) {
+    if (this.isSubmitClick) {
+        return true;
+    }
+
+    if (this.userForm.pristine) {
       return true;
     }