+++ src/app/core/@ngrx/products/products.actions.ts
@@ -1,0 +1,59 @@
+import { createAction, props } from '@ngrx/store';
+
+import { IProduct } from './../../../products/models/product.model';
+
+export const getProducts = createAction('[Products List Page (App)] GET_PRODUCTS');
+
+export const getProductsSuccess = createAction(
+  '[Get Products Effect] GET_PRODUCTS_SUCCEESS',
+  props<{ products: IProduct[] }>()
+);
+export const getProductsError = createAction(
+  '[Get Products Effect] GET_PRODUCTS_ERROR',
+  props<{ error: Error | string }>()
+);
+
+export const createProduct = createAction(
+  '[Admin Product Form Page] CREATE_PRODUCT',
+  props<{ product: IProduct }>()
+);
+
+export const createProductSuccess = createAction(
+  '[Create Product Effect] CREATE_PRODUCT_SUCCESS',
+  props<{ product: IProduct }>()
+);
+
+export const createProductError = createAction(
+  '[Create Product Effect] CREATE_PRODUCT_ERROR',
+  props<{ error: Error | string }>()
+);
+
+export const updateProduct = createAction(
+  '[Admin Product Form Page] UPDATE_PRODUCT',
+  props<{ product: IProduct }>()
+);
+
+export const updateProductSuccess = createAction(
+  '[Update Product Effect] UPDATE_PRODUCT_SUCCESS',
+  props<{ product: IProduct }>()
+);
+
+export const updateProductError = createAction(
+  '[Update Product Effect] UPDATE_PRODUCT_ERROR',
+  props<{ error: Error | string }>()
+);
+
+export const deleteProduct = createAction(
+  '[Admin Products List Page] DELETE_PRODUCT',
+  props<{ product: IProduct }>()
+);
+
+export const deleteProductSuccess = createAction(
+  '[Delete Product Effect] DELETE_PRODUCT_SUCCESS',
+  props<{ product: IProduct }>()
+);
+
+export const deleteProductError = createAction(
+  '[Delete Product Effect] DELETE_PRODUCT_ERROR',
+  props<{ error: Error | string }>()
+);