+++ src/app/core/@ngrx/root-store.module.ts
@@ -1,0 +1,35 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+// @NgRx
+import { StoreModule } from '@ngrx/store';
+import { ProductsStoreModule } from './products/products-store.module';
+
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { environment } from './../../../environments/environment';
+
+import { EffectsModule } from '@ngrx/effects';
+import { CustomSerializer, RouterEffects, routerReducers } from './router';
+import { RouterState, StoreRouterConnectingModule } from '@ngrx/router-store';
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    StoreModule.forRoot(routerReducers, {
+      runtimeChecks: {
+        strictStateSerializability: true,
+        strictActionWithinNgZone: true,
+        strictActionTypeUniqueness: true
+      }
+    }),
+    StoreRouterConnectingModule.forRoot({
+      stateKey: 'router',
+      routerState: RouterState.Minimal,
+      serializer: CustomSerializer
+    }),
+    EffectsModule.forRoot([RouterEffects]),
+    !environment.production ? StoreDevtoolsModule.instrument() : [],
+    ProductsStoreModule,
+  ]
+})
+export class RootStoreModule { }