+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,8 +1,8 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { CartService } from 'src/app/cart/services/cart.service';
-import { Product } from '../../models/product.model';
-import { ProductsService } from '../../services/products.service';
+import { ProductsFacade } from 'src/app/core/@ngrx/products';
+import { IProduct, Product } from '../../models/product.model';
 
 @Component({
   selector: 'app-product-list',
@@ -10,15 +10,19 @@ import { ProductsService } from '../../services/products.service';
   styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent implements OnInit {
-  products: Observable<Array<Product>>;
+  products$: Observable<ReadonlyArray<IProduct>>;
+  productsError$: Observable<Error | string>;
 
   constructor(
-    private productsService: ProductsService,
+    private productsFacade: ProductsFacade,
     private cartService: CartService
   ) {}
 
   ngOnInit(): void {
-    this.products = this.productsService.getProducts();
+    this.products$ = this.productsFacade.products$;
+    this.productsError$ = this.productsFacade.productsError$;
+
+    this.productsFacade.loadProducts();
   }
 
   onBuyProduct(product: Product): void {