+++ changelog.md
@@ -78,3 +78,10 @@ Removed obsolet comments
 
 31.03.2021
 Fixed bug with empty app-settings.json
+
+05.04.2021
+Added ngrx; Added Products store
+
+06.04.2021
+Connect Products store with components via facade
+Added Router store